Moved template definition to header file
Template definitions should not be in source files.
